requirements.txt
mysql-connector-python
Flask
flask_login
Flask-Toastr